+++ src/verticalaligner.cpp
@@ -85,6 +85,24 @@ StaffAlignment *SystemAligner::GetStaffAlignmentForStaffN(int staffN) const
     return NULL;
 }
 
+void SystemAligner::FindAllPositionerPointingTo(ArrayOfFloatingPositioners *positioners, FloatingObject *object)
+{
+    assert(positioners);
+
+    positioners->clear();
+
+    StaffAlignment *alignment = NULL;
+    int i;
+    for (i = 0; i < this->GetChildCount(); ++i) {
+        alignment = dynamic_cast<StaffAlignment *>(m_children.at(i));
+        assert(alignment);
+        FloatingPositioner *positioner = alignment->GetCorrespFloatingPositioner(object);
+        if (positioner && (positioner->GetObject() == object)) {
+            positioners->push_back(positioner);
+        }
+    }
+}
+
 //----------------------------------------------------------------------------
 // StaffAlignment
 //----------------------------------------------------------------------------